refactored blog posts page
