+++ src/data/posts.ts
@@ -1,0 +1,25 @@
+import { PostLinkProps } from '../pages/BlogpostList';
+
+export const POSTS: PostLinkProps[] = [
+    {
+        href: '/blog/first-post',
+        heading: 'Tech Diary #1: Getting A Space For The Blog',
+        description: 'Setting up a domain name and a hosting for the site',
+        date: 'sometime soon',
+        tags: ['tech_diary']
+    },
+    {
+        href: '/blog/second-post',
+        heading: 'Tech Diary #2: Dockerize The Site!',
+        description: 'Wrapping it all in Docker',
+        date: 'sometime soon',
+        tags: ['tech_diary']
+    },
+    {
+        href: '/blog/third-post',
+        heading: 'Tech Diary #3: Adding React, TS And Bundling',
+        description: 'Setting up a domain name and a hosting for the site',
+        date: 'sometime soon',
+        tags: ['tech_diary']
+    }
+];
